rootdir: disable core control driver
We are using cpusets and core control to manage the cores which is confusing things.
Disable core control because performance seems a lot better with cpusets, mainly because all cores are enabled all the time but I have noticed an improvement in battery life too.

Also use angler cpu architecture.

Change-Id: Ia9373b5f2f25747f8d6ede384f5afd8ac60744f7